feat: pre-commit added with ruff and mypy

+++ back/src/cube_analyzer/cardpool.py
@@ -2,7 +2,7 @@ import pandas as pd
 from src.database_manager.data_cleaner import remove_doublon, format_oracle_text
 
 
-def color_proportion(pool_cards: pd.DataFrame) -> dict[str, int]:
+def color_proportion(pool_cards: pd.DataFrame) -> dict[str, float]:
     dict_color = {}
     size = pool_cards.shape[0]
 
@@ -43,7 +43,7 @@ def esperance_to_find_keyword_by_booster(
         "rare": 7 / 8,
         "mythic": 1 / 8,
     }
-    esperance = 0
+    esperance = 0.0
     for rarity in dict_rarity_by_booster.keys():
         esperance += (
             dict_rarity_by_booster[rarity]
